frameclock: fix variable type
We count something, so use a size.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/10108>

+++ gdk/gdkframeclock.c
@@ -107,7 +107,7 @@ struct _GdkFrameClockPrivate
   uint64_t latest_presentation_time;
   uint64_t latest_refresh_interval;
 
-  int n_started;
+  gsize n_started;
   gsize n_updating;
 };
